+++ Snippets/svg2glif.py
@@ -1,8 +1,8 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 """ Convert SVG paths to UFO glyphs. """
 
 
-__requires__ = ["FontTools", "ufoLib"]
+__requires__ = ["fontTools"]
 
 from fontTools.misc.py23 import SimpleNamespace
 from fontTools.svgLib import SVGPath